update api to check if user bidded

+++ BackEnd/src/main/java/swp/koi/service/invoiceService/InvoiceService.java
@@ -18,7 +18,7 @@ public interface InvoiceService {
 
     List<Invoice> getAllInvoicesForAuctionWinner();
 
-    Invoice updateInvoiceAddress(double kilometer, int invoiceId, String address) throws UnsupportedEncodingException;
+    Invoice updateInvoiceAddress(float kilometer, int invoiceId, String address) throws UnsupportedEncodingException;
 
     Invoice getInvoiceForSpecificLot(int lotId);
